feat: single setup file for experiments
