+++ examples/config/paper/many_well.yaml
@@ -1,0 +1,49 @@
+# Currently with fake values so we can run locally with ease
+target:
+  dim: 4
+
+flow:
+  layer_nodes_per_dim: 10
+  n_layers: 10
+
+fab:
+  loss_type: alpha_2_div
+  transition_operator:
+    type: hmc
+    n_inner_steps: 5
+  n_intermediate_distributions: 2
+
+
+training:
+  seed: 0
+  lr: 2e-4
+  batch_size: 10
+  n_iterations: 100
+  use_gpu: true
+  use_64_bit: false
+  use_buffer: true # below config fields are all for use_buffer = True
+  prioritised_buffer: true
+  n_batches_buffer_sampling: 2
+  buffer_temp: 1.0 # rate that we weight new experience over old
+  maximum_buffer_length: 1000
+  min_buffer_length: 20 # heuristic: set this to n_batches_buffer_sampling*batch_size*10
+  log_w_clip_frac: null # null for no clipping, for non-prioritised replay
+  max_grad_norm: 10 # null for no clipping
+  w_adjust_max_clip: 10.0 # clipping of weight adjustment factor for prioritised replay
+
+
+evaluation:
+  n_plots: 10 # number of times we visualise the model throughout training.
+  n_eval: 10 # for calculating metrics of flow w.r.t target.
+  eval_batch_size: 200 # must be a multiple of inner batch size
+  n_checkpoints: 10 # number of model checkpoints saved
+  save_path:  results/many_well32/
+
+
+logger:
+  pandas_logger:
+    save_period: 100 # how often to save the pandas dataframe as a csv
+#  wandb:
+#    name: ManyWell32
+#    tags: [alpha_2_loss,ManyWell32]
+
