+++ examples/setup_run.py
@@ -1,21 +1,28 @@
-from typing import Union
-import matplotlib.pyplot as plt
-import torch
+from typing import Union, Callable, Optional, List
 import os
 import pathlib
-import hydra
 import wandb
 from omegaconf import DictConfig
 
 from datetime import datetime
 
 
-from fab import FABModel, HamiltoneanMonteCarlo, Trainer, Metropolis, BufferTrainer, PrioritisedBufferTrainer
+from fab import Trainer, BufferTrainer, PrioritisedBufferTrainer
+from fab.target_distributions.base import TargetDistribution
 from fab.utils.logging import PandasLogger, WandbLogger, Logger
 from fab.utils.replay_buffer import ReplayBuffer
-from fab.utils.prioritised_replay_buffer import PrioritisedReplayBuffer
-from fab.utils.plotting import plot_contours, plot_marginal_pair
+import matplotlib.pyplot as plt
+import torch
+
+from fab import FABModel, HamiltoneanMonteCarlo, Metropolis
 from examples.make_flow import make_wrapped_normflowdist
+
+from fab.utils.prioritised_replay_buffer import PrioritisedReplayBuffer
+
+
+Plotter = Callable[[FABModel], List[plt.Figure]]
+SetupPlotterFn = Callable[[DictConfig, TargetDistribution,
+                           Optional[Union[ReplayBuffer, PrioritisedReplayBuffer]]], Plotter]
 
 
 def setup_logger(cfg: DictConfig, save_path: str) -> Logger:
@@ -59,9 +66,9 @@ def setup_buffer(cfg: DictConfig, fab_model: FABModel) -> Union[ReplayBuffer,
     return buffer
 
 
-
-
-def _run(cfg: DictConfig):
+def setup_trainer_and_run(cfg: DictConfig, setup_plotter: SetupPlotterFn,
+                          target: TargetDistribution):
+    """Create and trainer and run."""
     dim = cfg.target.dim  # applies to flow and target
     if cfg.training.use_64_bit:
         torch.set_default_dtype(torch.float64)
@@ -77,10 +84,6 @@ def _run(cfg: DictConfig):
 
     with open(save_path + "config.txt", "w") as file:
         file.write(str(cfg))
-    from fab.target_distributions.many_well import ManyWellEnergy
-    assert dim % 2 == 0
-    target = ManyWellEnergy(dim, a=-0.5, b=-6)
-    plotting_bounds = (-3, 3)
 
     flow = make_wrapped_normflowdist(dim, n_flow_layers=cfg.flow.n_layers,
                                      layer_nodes_per_dim=cfg.flow.layer_nodes_per_dim)
@@ -121,51 +124,10 @@ def _run(cfg: DictConfig):
     # Create buffer if needed
     if cfg.training.use_buffer is True:
         buffer = setup_buffer(cfg, fab_model)
+    else:
+        buffer = None
 
-
-    def plot(fab_model, n_samples: int = cfg.training.batch_size, dim: int = dim):
-        n_rows = dim // 2
-        if cfg.training.prioritised_buffer is True:
-            fig, axs = plt.subplots(dim // 2, 3, sharex=True, sharey=True, figsize=(10, n_rows * 3))
-        else:
-            fig, axs = plt.subplots(dim // 2, 2, sharex=True, sharey=True, figsize=(10, n_rows * 3))
-
-        samples_flow = fab_model.flow.sample((n_samples,))
-        samples_ais = fab_model.annealed_importance_sampler.sample_and_log_weights(n_samples,
-                                                                                   logging=False)[0]
-        if cfg.training.prioritised_buffer is True:
-            samples_buffer = buffer.sample(n_samples)[0]
-
-        for i in range(n_rows):
-            plot_contours(target.log_prob_2D, bounds=plotting_bounds, ax=axs[i, 0])
-            plot_contours(target.log_prob_2D, bounds=plotting_bounds, ax=axs[i, 1])
-
-            # plot flow samples
-            plot_marginal_pair(samples_flow, ax=axs[i, 0], bounds=plotting_bounds,
-                               marginal_dims=(i * 2, i * 2 + 1))
-            axs[i, 0].set_xlabel(f"dim {i * 2}")
-            axs[i, 0].set_ylabel(f"dim {i * 2 + 1}")
-
-            # plot ais samples
-            plot_marginal_pair(samples_ais, ax=axs[i, 1], bounds=plotting_bounds,
-                               marginal_dims=(i * 2, i * 2 + 1))
-            axs[i, 1].set_xlabel(f"dim {i * 2}")
-            axs[i, 1].set_ylabel(f"dim {i * 2 + 1}")
-
-            if cfg.training.prioritised_buffer is True:
-                plot_contours(target.log_prob_2D, bounds=plotting_bounds, ax=axs[i, 2])
-                plot_marginal_pair(samples_buffer, ax=axs[i, 2], bounds=plotting_bounds,
-                                   marginal_dims=(i * 2, i * 2 + 1))
-                axs[i, 2].set_xlabel(f"dim {i * 2}")
-                axs[i, 2].set_ylabel(f"dim {i * 2 + 1}")
-
-            plt.tight_layout()
-        axs[0, 1].set_title("ais samples")
-        axs[0, 0].set_title("flow samples")
-        if cfg.training.use_buffer is True:
-            axs[0, 2].set_title("buffer samples")
-        plt.show()
-        return [fig]
+    plot = setup_plotter(cfg, target, buffer)
 
     # Create trainer
     if cfg.training.use_buffer is False:
@@ -174,16 +136,6 @@ def _run(cfg: DictConfig):
                           max_gradient_norm=cfg.training.max_grad_norm
                           )
     elif cfg.training.prioritised_buffer is False:
-        def initial_sampler():
-            # used to fill the replay buffer up to its minimum size
-            x, log_w = fab_model.annealed_importance_sampler.sample_and_log_weights(
-                    cfg.training.batch_size, logging=False)
-            return x, log_w
-
-        buffer = ReplayBuffer(dim=dim, max_length=cfg.training.maximum_buffer_length,
-                              min_sample_length=cfg.training.min_buffer_length,
-                              initial_sampler=initial_sampler,
-                              temperature=cfg.training.buffer_temp)
         trainer = BufferTrainer(model=fab_model, optimizer=optimizer, logger=logger, plot=plot,
                           optim_schedular=scheduler, save_path=save_path,
                                 buffer=buffer,
@@ -192,33 +144,16 @@ def _run(cfg: DictConfig):
                                 max_gradient_norm=cfg.training.max_grad_norm
                                 )
     else:
-        # buffer
-        def initial_sampler():
-            x, log_w = fab_model.annealed_importance_sampler.sample_and_log_weights(
-                cfg.training.batch_size, logging=False)
-            log_q_x = fab_model.flow.log_prob(x).detach()
-            return x, log_w, log_q_x
-
-        buffer = PrioritisedReplayBuffer(dim=dim, max_length=cfg.training.maximum_buffer_length,
-                              min_sample_length=cfg.training.min_buffer_length,
-                              initial_sampler=initial_sampler)
-
-        trainer = PrioritisedBufferTrainer(model=fab_model, optimizer=optimizer, logger=logger, plot=plot,
+        trainer = PrioritisedBufferTrainer(model=fab_model, optimizer=optimizer, logger=logger,
+                                           plot=plot,
                           optim_schedular=scheduler, save_path=save_path,
                                 buffer=buffer,
                                 n_batches_buffer_sampling=cfg.training.n_batches_buffer_sampling,
                                 max_gradient_norm=cfg.training.max_grad_norm,
                                 w_adjust_max_clip=cfg.training.w_adjust_max_clip
                                 )
+
     trainer.run(n_iterations=cfg.training.n_iterations, batch_size=cfg.training.batch_size,
                 n_plot=cfg.evaluation.n_plots,
                 n_eval=cfg.evaluation.n_eval, eval_batch_size=cfg.evaluation.eval_batch_size,
                 save=True, n_checkpoints=cfg.evaluation.n_checkpoints)
-
-
-@hydra.main(config_path="./config", config_name="many_well.yaml")
-def run(cfg: DictConfig):
-    _run(cfg)
-
-if __name__ == '__main__':
-    run()
